requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
pymongo>=4.6.0
jira>=3.5.0
atlassian-python-api>=3.41.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
google-cloud-storage>=2.14.0
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
